feat: rename to require-auth
