+++ src/middleware/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import requireAuth from './requireAuth';
+import requireAuth from './require-auth';
 
 export {
   requireAuth,
